% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{geo_data}
\alias{geo_data}
\title{Download format gene expression data from Gene Expression Omnibus}
\usage{
geo_data(dir = "./data-raw", gse = c("GSE202295"))
}
\arguments{
\item{dir}{Directory for downloading temporary files. Defaults to "data-raw"}

\item{gse}{GSE accession number}
}
\description{
Currently, a data sets from Pillon et al. is made
available for the user
after downloading from the Gene Expression Omnibus. The dataset can be used
for replicating analyses in seqwrap vignette and paper.
}
\references{
Pillon NJ, Smith JAB, Alm PS, Chibalin AV et al.
Distinctive exercise-induced inflammatory response and exerkine
induction in skeletal muscle of people with type 2 diabetes.
\emph{Sci Adv} 2022 Sep 9;8(36):eabo3192. PMID: 36070371
}
