% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shetty_transcript}
\alias{shetty_transcript}
\title{Transcript-level data from Shetty et al. 2025}
\format{
A data frame with 387944 rows and 71 variables:
\describe{
\item{gene_id}{Gene id Ensembl }
\item{seq_sample_id}{Sequencing sample ID, e.g., GSM7330944 corresponding to seq_sample_id in the meta data.}
}
}
\usage{
shetty_transcript
}
\description{
Transcript-level data from Shetty et al. 2025
}
\references{
Shetty AC, Sivinski J, Cornell J, McCracken C et al. Peripheral
blood transcriptomic profiling of molecular mechanisms commonly
regulated by binge drinking and placebo effects.
\emph{Sci Rep} 2024 May 10;14(1):10733. PMID: 38730024
}
\keyword{datasets}
